% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_sdgs}
\alias{available_sdgs}
\title{available_sdgs}
\usage{
available_sdgs()
}
\value{
prints all the SDG names
}
\description{
Lists all the available SDGs computable through `generate_sdg_report()` 
function. They can be used when specifing the variable `sdgs` in the previous
function.
}
