% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prj.R
\name{data_query}
\alias{data_query}
\title{data_query}
\usage{
data_query(type, db_path, db_name, prj_name, scenarios)
}
\arguments{
\item{type}{query name}

\item{db_path}{database path}

\item{db_name}{database name}

\item{prj_name}{project name}

\item{scenarios}{scenarios to be considered}
}
\value{
dataframe with the specified query information
}
\description{
Aux. function to load heavy queries
}
