% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_prj.R
\name{.create_prj}
\alias{.create_prj}
\title{.create_prj}
\usage{
.create_prj(
  db_name,
  db_path,
  prj_name = NULL,
  output_name = NULL,
  desired_scen = NULL,
  required_queries = "All",
  include_land_query = TRUE,
  include_nonco2_query = TRUE
)
}
\arguments{
\item{db_name}{name of the database. It will The extension will be automatically 
added if not present}

\item{db_path}{run directory containing the GCAM database}

\item{prj_name}{name of the project. If NULL, it will be the default option, 
i.e., the database name. Otherwise specify}

\item{output_name}{name of the output file. When processing multiple projects 
(`prj_name`), this specifies the filename for the SDG outputs saved in 
the 'output' directory. Defaults to the first `prj_name`.}

\item{desired_scen}{desired scenarios. If NULL, all the scenarios present in 
the database will be considered}

\item{required_queries}{required queries to be loaded in the project. If 'All'
(default), all the necessary queries to compute the required SDGs will be considered}

\item{include_land_query}{whether to merge in the "detailed land allocation"
query (needed for the SDG15 land indicator). Default TRUE; run() sets
this to FALSE when SDG15 isn't among the requested indicators, to avoid
the extra extraction cost.}

\item{include_nonco2_query}{whether to merge in the "nonCO2 emissions by
sector" query (needed for the SDG3 health indicator via rfasst).
Default TRUE; run() sets this to FALSE when SDG3 isn't requested, since
this query is chunked and comparatively slow to extract.}
}
\value{
create the specified project
}
\description{
Function to create a GCAM project provided a database and a queries file
}
