% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.gather_sdgs}
\alias{.gather_sdgs}
\title{.gather_sdgs}
\usage{
.gather_sdgs(result)
}
\arguments{
\item{result}{list containing the calculated SDG indicators.}
}
\value{
standardised dataset with all the SDGs
}
\description{
Gathers all the SDG individual indicators and binds them into a single dataset.
}
