% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_food_weights}
\alias{.get_food_weights}
\title{.get_food_weights}
\usage{
.get_food_weights()
}
\value{
`food_weights` and `food_wld_weights` global variables.
}
\description{
Get food items weighted by demand. By region and global.
Source: adapted from gcamreport pkg.
}
\keyword{internal}
