% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdg.R
\name{.make_sdg_figures}
\alias{.make_sdg_figures}
\title{.make_sdg_figures}
\usage{
.make_sdg_figures(result, output_name)
}
\arguments{
\item{result}{list of reported SDGs}

\item{output_name}{name of the output file. When processing multiple projects 
(`prj_name`), this specifies the filename for the SDG outputs saved in 
the 'output' directory. Defaults to the first `prj_name`.}
}
\description{
Generate a basic figure for each raw (non-diffed) SDG indicator in
`result`, saved in `output/figures`.
}
\keyword{internal}
