% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.transform_to_xml}
\alias{.transform_to_xml}
\title{.transform_to_xml}
\usage{
.transform_to_xml(parsed_queries_list)
}
\arguments{
\item{parsed_queries_list}{List of parsed queries.}
}
\value{
XML document generated from the provided queries list.
}
\description{
Converts a list of parsed queries into an XML document.
Source: adapted from gcamreport
}
\keyword{internal}
