% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_subsector}
\alias{food_subsector}
\title{food_subsector}
\format{
.csv
}
\source{
local
}
\usage{
food_subsector
}
\description{
food subsector mapping
}
\examples{
\dontrun{
library(gcamsdg)
gcamsdg::food_subsector
}
}
\keyword{datasets}
