% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg1_gdp.R
\name{get_sdg1_gdp}
\alias{get_sdg1_gdp}
\title{get_sdg1_gdp}
\usage{
get_sdg1_gdp(prj_f, output_name, saveOutput = T, makeFigures = F)
}
\arguments{
\item{prj_f}{uploaded project file}

\item{output_name}{output file name, used to tag the saved output file in 
the 'output' directory.}

\item{saveOutput}{save the produced output}

\item{makeFigures}{generate and save graphical representation/s of the output}
}
\value{
data frame with GDP per capita (PPP) by region, scenario and year
}
\description{
Extract GDP per capita (PPP) by region, used as the SDG 1 (Poverty) economy
metric and as an input to SDG2's food-basket-bill-as-%-of-GDP calculation.
}
