% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prj.R
\name{load_prj}
\alias{load_prj}
\title{load_prj}
\usage{
load_prj(prj_path, prj_name)
}
\arguments{
\item{prj_path}{directory containing the project file}

\item{prj_name}{project file name}
}
\value{
the loaded rgcam project
}
\description{
Load a previously saved rgcam project file.
}
