% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_land}
\alias{query_land}
\title{query_land}
\format{
.csv
}
\source{
local
}
\usage{
query_land
}
\description{
query file to report the SDGs 15
}
\examples{
\dontrun{
library(gcamsdg)
gcamsdg::query_land
}
}
\keyword{datasets}
