% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_nonCO2}
\alias{query_nonCO2}
\title{query_nonCO2}
\format{
.csv
}
\source{
local
}
\usage{
query_nonCO2
}
\description{
query file to report SDGs 3 (necessary for rfasst)
}
\examples{
\dontrun{
library(gcamsdg)
gcamsdg::query_nonCO2
}
}
\keyword{datasets}
