% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg2_food_basket_bill.R
\name{get_sdg2_food_basket_bill}
\alias{get_sdg2_food_basket_bill}
\title{get_sdg2_food_basket_bill}
\usage{
get_sdg2_food_basket_bill(prj, output_name, saveOutput = T, makeFigures = F)
}
\arguments{
\item{prj}{uploaded project file}

\item{output_name}{output file name, used to tag the saved output file in 
the 'output' directory.}

\item{saveOutput}{save the produced output}

\item{makeFigures}{generate and save graphical representation/s of the output}
}
\value{
data frame with the global food basket bill (% GDP) by scenario and year
}
\description{
Compute SDG 2 (Zero Hunger) as the per-capita food basket bill, expressed
as a percentage of GDP, weighted globally by population.
}
