% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg3_health.R
\name{get_sdg3_health}
\alias{get_sdg3_health}
\title{get_sdg3_health}
\usage{
get_sdg3_health(prj, output_name, saveOutput = T, makeFigures = F)
}
\arguments{
\item{prj}{uploaded project file}

\item{output_name}{output file name, used to tag the saved output file in 
the 'output' directory.}

\item{saveOutput}{save the produced output}

\item{makeFigures}{generate and save graphical representation/s of the output}
}
\value{
data frame with mortalities by scenario, GCAM region and year
}
\description{
Compute SDG 3 (Health) as premature mortalities attributable to long-term
exposure to PM2.5 and O3, using rfasst, downscaled to country level by
population share and re-aggregated to GCAM region.
}
